requirements.txt
# Core dependencies
langgraph>=0.0.40
langchain>=0.1.0
langchain-openai>=0.1.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
faiss-cpu>=1.7.0
streamlit>=1.28.0
redis>=5.0.0

# FastAPI and server dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Database and async
asyncpg>=0.29.0
sqlalchemy>=2.0.0

# HTTP client
httpx>=0.25.0
requests>=2.31.0
python-multipart>=0.0.20